Rework segmentation between darwin/linux
