+++ cmd/starter/c/include/capability.h
