+++ internal/app/singularity/instance_stop.go
@@ -3,7 +3,7 @@
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
 
-// +build singularity_runtime
+// +build linux
 
 package cli
