+++ internal/app/singularity/oci.go
@@ -3,6 +3,8 @@
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
 
+// +build linux
+
 package cli
 
 import (
