--- internal/pkg/runtime/c/lib/util/util.mconf
@@ -1,5 +1,0 @@
-name := libutil
-lib := util
-csrc :=	internal/pkg/runtime/c/lib/util/capability.c
-	internal/pkg/runtime/c/lib/util/message.c
-	internal/pkg/runtime/c/lib/util/setns.c
