+++ internal/pkg/runtime/engines/singularity/config/data/singularity.conf
