+++ internal/pkg/runtime/engines/singularity/engine.go
@@ -7,12 +7,13 @@ package singularity
 
 import (
 	"github.com/sylabs/singularity/internal/pkg/runtime/engines/config"
+	singularityConfig "github.com/sylabs/singularity/internal/pkg/runtime/engines/singularity/config"
 )
 
 // EngineOperations describes a runtime engine
 type EngineOperations struct {
-	CommonConfig *config.Common `json:"-"`
-	EngineConfig *EngineConfig  `json:"engineConfig"`
+	CommonConfig *config.Common                  `json:"-"`
+	EngineConfig *singularityConfig.EngineConfig `json:"engineConfig"`
 }
 
 // InitConfig stores the pointer to config.Common
