+++ mlocal/frags/Makefile.stub
@@ -1,11 +1,12 @@
-libruntime := $(BUILDDIR)/lib/libruntime.a
-
 starter_OBJ := $(shell $(SOURCEDIR)/makeit/gengodep $(SOURCEDIR)/cmd/starter/main.go)
 
 starter := $(BUILDDIR)/cmd/starter/c/starter
 starter_INSTALL := $(DESTDIR)$(LIBEXECDIR)/singularity/bin/starter
 starter_suid_INSTALL := $(DESTDIR)$(LIBEXECDIR)/singularity/bin/starter-suid
-starter_SOURCE := $(SOURCEDIR)/cmd/starter/c/starter.c
+starter_SOURCE := $(SOURCEDIR)/cmd/starter/c/starter.c \
+                  $(SOURCEDIR)/cmd/starter/c/util/capability.c \
+                  $(SOURCEDIR)/cmd/starter/c/util/message.c \
+                  $(SOURCEDIR)/cmd/starter/c/util/setns.c
 
 singularity := $(BUILDDIR)/singularity
 singularity_INSTALL := $(DESTDIR)$(EXECPREFIX)/bin/singularity
@@ -100,13 +101,13 @@ INSTALLFILES := $(singularity_INSTALL) $(bash_completion_INSTALL)
 
 ifeq ($(BUILD_RUNTIME),1)
 
-CLEANFILES += $(libruntime) $(starter) $(cni_plugins_EXECUTABLES)
+CLEANFILES += $(starter) $(cni_plugins_EXECUTABLES)
 INSTALLFILES += $(starter_INSTALL) $(starter_suid_INSTALL) $(sessiondir) \
 	$(config_INSTALL) $(dist_bin_SCRIPTS_INSTALL) $(capability_JSON) $(syecl_config_INSTALL) \
 	$(actions_INSTALL) $(cni_plugins_INSTALL) $(cni_config_INSTALL) $(seccomp_profile_INSTALL) \
 	$(NVIDIA_liblist_INSTALL) $(cgroups_config_INSTALL)
 
-all: cscope collect $(libruntime) $(go_BIN) $(config) $(starter) $(bash_completion) cniplugins
+all: cscope collect $(go_BIN) $(config) $(starter) $(bash_completion) cniplugins
 
 else # build_runtime
 
@@ -137,24 +138,19 @@ $(cni_config_INSTALL): $(cni_config_LIST)
 	$(V)install -d $(cni_config_INSTALL)
 	$(V)install -m 0644 $? $@
 
-$(libruntime): $(libutil_OBJ)
-	@echo " AR" $@
-	$(V)mkdir -p $(@D)
-	$(V)$(AR) rcs $@ $(libutil_OBJ)
-
 $(BUILDDIR)/.clean-starter: $(starter_SOURCE)
 	@echo " GO clean -cache"
 	$(V)(go clean -cache 2>/dev/null || true)
 	$(V)touch $@
 
 # starter
-$(starter): $(BUILDDIR)/.clean-starter $(go_OBJ) $(starter_OBJ) $(libruntime)
+$(starter): $(BUILDDIR)/.clean-starter $(go_OBJ) $(starter_OBJ)
 	@echo " GO" $@
 	$(V)go build $(GO_BUILDMODE) -tags "$(GO_TAGS)" $(GO_LDFLAGS) -o $@ \
 		$(SOURCEDIR)/cmd/starter/main.go
 
 # singularity
-$(singularity): $(go_OBJ) $(libruntime) $(singularity_OBJ)
+$(singularity): $(go_OBJ) $(singularity_OBJ)
 	@echo " GO" $@
 	@echo "    [+] GO_TAGS" \"$(GO_TAGS)\"
 	$(V)go build $(GO_BUILDMODE) -tags "$(GO_TAGS)" $(GO_LDFLAGS) -o $(BUILDDIR)/singularity $(SOURCEDIR)/cmd/singularity/cli.go
@@ -164,12 +160,13 @@ $(singularity_INSTALL): $(singularity)
 	$(V)install -m 0755 $(singularity) $(singularity_INSTALL) # set cp to install
 
 # config.go
-$(go_OBJ): $(libruntime) $(BUILDDIR)/config.h
+$(go_OBJ): $(BUILDDIR)/config.h
 	$(V)rm -f $(go_OBJ)
 	$(V)export BUILDDIR=$(BUILDDIR_ABSPATH) && cd $(SOURCEDIR)/internal/pkg/buildcfg && go generate
 
 # bash_completion file
 $(bash_completion): $(go_OBJ)
+	@echo " GEN" $@
 	$(V)rm -f $@
 	$(V)mkdir -p $(@D)
 	$(V)go run -tags "$(GO_TAGS)" $(SOURCEDIR)/etc/bash_completion.d/bash_completion.go $@
@@ -202,9 +199,10 @@ $(dist_bin_SCRIPTS_INSTALL): $(dist_bin_SCRIPTS)
 	$(V)install -d $(@D)
 	$(V)install -m 0755 $(dist_bin_SCRIPTS) $(dist_bin_SCRIPTS_INSTALL)
 
-$(config): $(go_OBJ) $(SOURCEDIR)/etc/conf/gen.go $(SOURCEDIR)/internal/pkg/runtime/engines/singularity/data/singularity.conf $(SOURCEDIR)/internal/pkg/runtime/engines/singularity/config.go
+$(config): $(go_OBJ) $(SOURCEDIR)/etc/conf/gen.go $(SOURCEDIR)/internal/pkg/runtime/engines/singularity/config/data/singularity.conf $(SOURCEDIR)/internal/pkg/runtime/engines/singularity/config/config.go
 	@echo " GEN $@`if [ -n "$(old_config)" ]; then echo " from $(old_config)"; fi`"
-	$(V)go run $(SOURCEDIR)/etc/conf/gen.go $(SOURCEDIR)/internal/pkg/runtime/engines/singularity/data/singularity.conf $(old_config) $(config)
+	$(V)go run $(SOURCEDIR)/etc/conf/gen.go \
+		$(SOURCEDIR)/internal/pkg/runtime/engines/singularity/config/data/singularity.conf $(old_config) $(config)
 
 $(config_INSTALL): $(config)
 	@echo " INSTALL" $@
