+++ mlocal/frags/go_runtime_opts.mk
@@ -1,8 +1,4 @@
-GO_TAGS += singularity_runtime
-
-CGO_CPPFLAGS := -I$(BUILDDIR) -I$(SOURCEDIR)/cmd/starter/c -I$(SOURCEDIR)/internal/pkg/runtime/c/lib
+CGO_CPPFLAGS += -I$(BUILDDIR) -I$(SOURCEDIR)/cmd/starter/c -I$(SOURCEDIR)/cmd/starter/c/include
 CGO_CPPFLAGS += -include $(BUILDDIR_ABSPATH)/config.h
-
-CGO_LDFLAGS := -L$(BUILDDIR_ABSPATH)/lib -L$(BUILDDIR) -lruntime
 
 export CGO_CPPFLAGS CGO_LDFLAGS
